fix cpp findings
