+++ Profiler/WindowsUtils.cpp
